Better support for addon soft errors.
