+++ hack/cmd/addon/main.go
@@ -24,6 +24,8 @@ var (
 	Log   = hub.Log
 )
 
+type SoftError = hub.SoftError
+
 //
 // main
 func main() {
